Lib: move template related to subnamespace
