+++ lib/Templates/Timber.php
@@ -6,7 +6,7 @@
  * @subpackage PixelsTheme
  */
 
-namespace Pixels\Theme;
+namespace Pixels\Theme\Templates;
 
 use Pixels\Theme\Twig\Context   as Context;
 use Pixels\Theme\Twig\Functions as Functions;
